+++ NEWS.md
@@ -2,6 +2,8 @@
 
 # development version
 
+- Add Mou 2024 aripiprazole ([doi:10.1177/20420986241303432](https://doi.org/10.1177/20420986241303432)) -- adults by CYP2D6 phenotype, with mild hepatic impairment and risperidone coadministration; a compartmental reduction of the published PK-Sim PBPK model.
+
 - Add Moein 2024 apitolisib translational PK/PD models ([doi:10.1007/s40268-024-00459-5](https://doi.org/10.1007/s40268-024-00459-5)) -- mouse (786-O renal cell adenocarcinoma xenograft) and adults with advanced solid tumors or non-Hodgkin's lymphoma.
 
 - Add Decrane 2023 oxyfluorfen ([doi:10.1016/j.crtox.2023.100138](https://doi.org/10.1016/j.crtox.2023.100138)) -- rat (Sprague-Dawley) and extrapolated human.